added /transfers endpoint

+++ src/main/java/misrraimsp/fourthrest/model/dto/ExpenseConverter.java
@@ -3,7 +3,6 @@ package misrraimsp.fourthrest.model.dto;
 import misrraimsp.fourthrest.model.Expense;
 import misrraimsp.fourthrest.model.Person;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class ExpenseConverter {
@@ -15,7 +14,7 @@ public class ExpenseConverter {
         dto.setPayerFirstName(expense.getPayer().getFirstName());
         dto.setPayerLastName(expense.getPayer().getLastName());
         dto.setDescription(expense.getDescription());
-        dto.setAmount(expense.getAmount().doubleValue());
+        dto.setAmount(expense.getAmount());
         dto.setDate(expense.getDate().toString());
         return dto;
     }
@@ -26,7 +25,7 @@ public class ExpenseConverter {
         payer.setId(dto.getPayerId());
         expense.setPayer(payer);
         expense.setDescription(dto.getDescription());
-        expense.setAmount(BigDecimal.valueOf(dto.getAmount()));
+        expense.setAmount(dto.getAmount());
         expense.setDate(LocalDateTime.now());
         return expense;
     }
