+++ src/main/java/misrraimsp/fourthrest/model/dto/ExpenseDTO.java
@@ -2,20 +2,16 @@ package misrraimsp.fourthrest.model.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ExpenseDTO {
 
     private Long id;
-
     private Long payerId;
-
     private String payerFirstName;
-
     private String payerLastName;
-
-    private double amount;
-
+    private BigDecimal amount;
     private String description;
-
     private String date;
 }
