+++ src/main/java/misrraimsp/fourthrest/model/dto/PersonDTO.java
@@ -2,14 +2,13 @@ package misrraimsp.fourthrest.model.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PersonDTO {
 
     private Long id;
-
     private String firstName;
-
     private String lastName;
-
-    private double balance;
+    private BigDecimal balance;
 }
